requirements.txt
openai
qdrant_client
python-dotenv
streamlit